backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
motor==3.1.1
pymongo==4.3.3
python-dotenv==1.0.0
pydantic==2.4.2
python-multipart==0.0.6 